# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OTP request and validation service.",
        "title": "OTP Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/healthz": {
            "get": {
                "description": "Check whether the service process is running.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/healthhttp.HealthResponse"
                        }
                    }
                }
            }
        },
        "/otp/request": {
            "post": {
                "description": "Generate a new OTP for a user based on the configured active-OTP policy.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "otp"
                ],
                "summary": "Request OTP",
                "parameters": [
                    {
                        "description": "Request OTP payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/otphttp.RequestOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/otphttp.RequestOTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpadapter.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/httpadapter.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpadapter.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/otp/validate": {
            "post": {
                "description": "Validate an OTP for a user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "otp"
                ],
                "summary": "Validate OTP",
                "parameters": [
                    {
                        "description": "Validate OTP payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/otphttp.ValidateOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/otphttp.ValidateOTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpadapter.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpadapter.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "$ref": "#/definitions/httpadapter.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/httpadapter.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpadapter.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "healthhttp.HealthData": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "healthhttp.HealthResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/healthhttp.HealthData"
                },
                "meta": {
                    "$ref": "#/definitions/httpadapter.Meta"
                }
            }
        },
        "httpadapter.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "otp_invalid"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "OTP is invalid"
                }
            }
        },
        "httpadapter.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/httpadapter.ErrorBody"
                },
                "meta": {
                    "$ref": "#/definitions/httpadapter.Meta"
                }
            }
        },
        "httpadapter.Meta": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string",
                    "example": "5f4dcc3b-5aa7"
                }
            }
        },
        "otphttp.RequestOTPData": {
            "type": "object",
            "properties": {
                "otp": {
                    "type": "string",
                    "example": "61531"
                },
                "user_id": {
                    "type": "string",
                    "example": "Robert"
                }
            }
        },
        "otphttp.RequestOTPRequest": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string",
                    "example": "Robert"
                }
            }
        },
        "otphttp.RequestOTPResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/otphttp.RequestOTPData"
                },
                "meta": {
                    "$ref": "#/definitions/httpadapter.Meta"
                }
            }
        },
        "otphttp.ValidateOTPData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "OTP validated successfully."
                },
                "user_id": {
                    "type": "string",
                    "example": "Robert"
                }
            }
        },
        "otphttp.ValidateOTPRequest": {
            "type": "object",
            "properties": {
                "otp": {
                    "type": "string",
                    "example": "61531"
                },
                "user_id": {
                    "type": "string",
                    "example": "Robert"
                }
            }
        },
        "otphttp.ValidateOTPResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/otphttp.ValidateOTPData"
                },
                "meta": {
                    "$ref": "#/definitions/httpadapter.Meta"
                }
            }
        }
    }
}